Introducing the multi-language support #11
The application is now supporting a multilanguage version, including English, Français and Deutsch (and increasing).
It uses by default the application device settings but it can be change on the application.
If the language is unknown, English will be used.

The language settings are loaded meanwhile the splashscreen is displayed (see following commit).